Refactor : Move HintManagerHolder;
